+++ model/pack.go
@@ -33,9 +33,9 @@ type Pack struct {
 	Logo          *Attachment  `json:"logo" gorm:"polymorphic:Owner"`
 	Background    *Attachment  `json:"background" gorm:"polymorphic:Owner"`
 	Recommended   *Build       `json:"recommended"`
-	RecommendedID int          `json:"recommended_id" sql:"index"`
+	RecommendedID int64        `json:"recommended_id" sql:"index"`
 	Latest        *Build       `json:"latest"`
-	LatestID      int          `json:"latest_id" sql:"index"`
+	LatestID      int64        `json:"latest_id" sql:"index"`
 	Slug          string       `json:"slug" sql:"unique_index"`
 	Name          string       `json:"name" sql:"unique_index"`
 	Website       string       `json:"website"`
